delayed on/off
Since Cayenne doesn't have the flexibility to keep track
of anything other than the current state for notifications,
we do that here. There's a little delay for the on/off transitions. It's particularly important for the washing
machine because it goes through so many power changes
during a cycle. I'm hoping 5 minutes is enough to avoid
false positives for the washer turning off.